+++ src/libs/config/config.interface.ts
