+++ src/libs/random/utils.ts
@@ -1,4 +1,3 @@
 export const getRandomInt = () => Math.floor(Math.random() * Number.MAX_SAFE_INTEGER);
 
 export const getRandomArbitraryInt = (min: number, max: number) => Math.floor(Math.random() * (max - min) + min);
-
