+++ src/rest/config/config.interface.ts
@@ -1,0 +1,3 @@
+export interface Config<U> {
+  get<T extends keyof U>(key: T): U[T];
+}
