+++ src/rest/config/rest.config.ts
@@ -1,7 +1,7 @@
 import { config } from 'dotenv';
 import { inject, injectable } from 'inversify';
-import { Config } from './config.interface.js';
-import { Logger } from '../logger/index.js';
+import { Config } from '../../libs/config/index.js';
+import { Logger } from '../../libs/logger/index.js';
 import { configRestSchema, RestSchema } from './rest.schema.js';
 import { Component } from '../component.js';
